refactor/#6 : 회원 몸무게와 신장 실수형으로 타입 변경

+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/CreateMemberService.java
@@ -30,13 +30,13 @@ public class CreateMemberService implements CreateMemberUsecase {
         validateAge(command.age());
     }
 
-    private void validateHeight(Integer height) {
+    private void validateHeight(Double height) {
         if (height <= MINIMUM_HEIGHT) {
             throw new InvalidHeightException("height must be greater than 0");
         }
     }
 
-    private void validateWeight(Integer weight) {
+    private void validateWeight(Double weight) {
         if (weight <= MINIMUM_WEIGHT) {
             throw new InvalidWeightException("weight must be greater than 0");
         }
