+++ BE/exceed/src/test/java/com/gaebaljip/exceed/member/application/CreateMemberServiceTest.java
@@ -21,7 +21,7 @@ class CreateMemberServiceTest {
     @InjectMocks
     private CreateMemberService createMemberService;
 
-    private static CreateMemberCommand createMemberCommand(int height, int weight, int age) {
+    private static CreateMemberCommand createMemberCommand(double height, double weight, int age) {
         CreateMemberCommand command = CreateMemberCommand.builder()
                 .height(height)
                 .weight(weight)
@@ -31,7 +31,6 @@ class CreateMemberServiceTest {
                 .build();
         return command;
     }
-
     @Test
     @DisplayName("키가 0 이하일 경우 예외가 발생한다.")
     void createMember_weight() {
